move greyvest components to src root
