+++ src/BarChart.js
