+++ src/Modal.js
@@ -3,7 +3,7 @@ import _ from 'lodash/fp'
 import { observer } from 'mobx-react'
 import Portal from './Portal'
 import { openBinding } from './utils'
-import { expandProp } from '../utils/react'
+import { expandProp } from './utils/react'
 
 let Modal = ({ isOpen, onClose, children, style = {}, className = '' }) => (
   <Portal>
