+++ src/Popover.js
@@ -3,7 +3,7 @@ import _ from 'lodash/fp'
 import OutsideClickHandler from 'react-outside-click-handler'
 import { observer } from 'mobx-react'
 import { openBinding } from './utils'
-import { expandProp } from '../utils/react'
+import { expandProp } from './utils/react'
 
 // Simple popover
 let Popover = ({ isOpen, onClose, children, style }) =>
