+++ src/TextHighlight.js
