--- src/greyVest/index.js
@@ -1,41 +1,0 @@
-export Box from './Box'
-export Button from './Button'
-export ButtonGroup from './ButtonGroup'
-export ButtonRadio from './ButtonRadio'
-export Checkbox from './Checkbox'
-export CheckboxList from './CheckboxList'
-export ErrorList from './ErrorList'
-export Fonts from './Fonts'
-export GridItem from './GridItem'
-export Icon from './Icon'
-export LinkButton from './LinkButton'
-export DropdownItem from './DropdownItem'
-export Modal from './Modal'
-export PagerItem from './PagerItem'
-export RadioList from './RadioList'
-export Select from './Select'
-export Style from './Style'
-export * from './Tabs'
-export Table from './Table'
-export TabList from './TabList'
-export Textarea from './Textarea'
-export TextButton from './TextButton'
-export TextHighlight from './TextHighlight'
-export TextInput from './TextInput'
-
-// former layout components
-// ------------------------
-export Awaiter from './Awaiter'
-export BarChart from './BarChart'
-export Dynamic from './Dynamic'
-export ExpandableTable, { Column } from './ExpandableTable'
-export Flex from './Flex'
-export Grid from './Grid'
-export NestedPicker from './NestedPicker'
-export Popover from './Popover'
-export Portal from './Portal'
-export SpacedList from './SpacedList'
-export TagsInput from './TagsInput'
-export Tag from './Tag'
-export DateInput from './DateInput'
-export StripedLoader from './StripedLoader'
