+++ src/stories/Box.stories.js
@@ -1,6 +1,6 @@
 import React from 'react'
-import { Box } from '.'
-import decorator from './stories/decorator'
+import { Box } from '..'
+import decorator from './decorator'
 
 export default {
   title: 'Components|Box',
