+++ src/stories/Popover.stories.js
@@ -1,8 +1,8 @@
 import React from 'react'
 import F from 'futil'
 import { observable } from 'mobx'
-import { Button, Popover } from '.'
-import decorator from './stories/decorator'
+import { Button, Popover } from '..'
+import decorator from './decorator'
 
 export default {
   title: 'Components|Popover',
